move project card
